🏗️ Architect: Modularized Observation class
Refactored the 1,163-line `apts/observations.py` 'God Object' into a modular package `apts/observations/`.

- Decomposed monolithic `Observation` class using a MixIn pattern:
  - `WeatherAnalysisMixIn` in `weather.py` handles complex weather suitability calculations.
  - `PlottingMixIn` in `plotting.py` encapsulates extensive plotting delegation wrappers.
  - `Observation` base class in `base.py` manages core state and data access.
- Maintained public API compatibility via `__init__.py`.
- Updated unit tests (`tests/unit/test_moon_condition_optimization.py`, etc.) to correct mock paths after relocation.

This improves maintainability and readability by logically separating concerns while preserving backward compatibility.
