--- apts/observations.py
@@ -1,1163 +1,0 @@
-import html
-import logging
-import re
-from datetime import datetime, timedelta
-from importlib import resources
-from string import Template
-from typing import TYPE_CHECKING, List, Optional, cast
-
-if TYPE_CHECKING:
-    import matplotlib.figure
-
-import numpy as np
-import pandas as pd
-from skyfield.api import utc
-
-from apts.constants.plot import CoordinateSystem
-
-from . import plot as apts_plot
-from .conditions import Conditions
-from .constants.event_types import EventType
-from .events import AstronomicalEvents
-from .i18n import language_context
-from .objects.messier import Messier
-from .objects.ngc import NGC
-from .objects.solar_objects import SolarObjects
-from .objects.stars import Stars
-from .utils import Utils
-from .utils.planetary import get_moon_illumination
-
-logger = logging.getLogger(__name__)
-
-
-class Observation:
-    NOTIFICATION_TEMPLATE = resources.files("apts").joinpath(
-        "templates/notification.html.template"
-    )
-
-    def __init__(
-        self,
-        place,
-        equipment,
-        conditions=Conditions(),
-        target_date=None,
-        sun_observation=False,
-        start_time: Optional[datetime] = None,
-        end_time: Optional[datetime] = None,
-        limiting_magnitude: Optional[float] = None,
-    ):
-        self.place = place
-        self.equipment = equipment
-        self.conditions = conditions
-        self.sun_observation = sun_observation
-        self.limiting_magnitude = limiting_magnitude
-
-        # Initialize core attributes that depend on date calculations
-        self.effective_date = None
-        self.observation_local_time = None
-        self.start = None
-        self.stop = None
-        self.time_limit = None
-
-        if sun_observation and start_time and end_time:
-            self.start = start_time
-            self.stop = end_time
-            self.effective_date = (
-                start_time  # Use start_time as effective date for calculations
-            )
-            self.observation_local_time = (
-                start_time  # Use start_time as local observation time
-            )
-        elif target_date:
-            if self.sun_observation:
-                self.start = self.place.sunrise_time(target_date=target_date)
-                self.stop = self.place.sunset_time(target_date=target_date)
-            else:
-                self.start = self.place.sunset_time(
-                    target_date=target_date, twilight=self.conditions.twilight
-                )
-                if self.start:
-                    self.stop = self.place.sunrise_time(
-                        start_search_from=self.start, twilight=self.conditions.twilight
-                    )
-                else:
-                    self.stop = None
-
-            if self.start is None or self.stop is None:
-                logger.warning(
-                    f"Could not determine observation window for {self.place.name} "
-                    f"on {target_date} with twilight '{self.conditions.twilight.value}'. "
-                    "Sun may be always up or down."
-                )
-            else:
-                if self.conditions.start_time:
-                    if isinstance(self.conditions.start_time, str):
-                        parts = [int(v) for v in self.conditions.start_time.split(":")]
-                        h = parts[0]
-                        m = parts[1] if len(parts) > 1 else 0
-                        s = parts[2] if len(parts) > 2 else 0
-                        override_start_dt = self.start.replace(
-                            hour=h,
-                            minute=m,
-                            second=s,
-                        )
-                    else:
-                        # Assume it's a datetime/time object and take its time components
-                        override_start_dt = self.start.replace(
-                            hour=self.conditions.start_time.hour,
-                            minute=self.conditions.start_time.minute,
-                            second=self.conditions.start_time.second,
-                        )
-                    self.start = override_start_dt
-
-                self.effective_date = self.place.ts.utc(self.start)
-                self.observation_local_time = self.start
-
-        else:
-            # Legacy behavior: use place.date
-            self.effective_date = self.place.date
-            if sun_observation:
-                self.start = self.place.sunrise_time()
-                self.stop = self.place.sunset_time()
-            else:
-                self.start = self.place.sunset_time(twilight=self.conditions.twilight)
-                if self.start:
-                    self.stop = self.place.sunrise_time(
-                        start_search_from=self.start, twilight=self.conditions.twilight
-                    )
-                else:
-                    self.stop = None
-            self.observation_local_time = self.start
-
-        # Normalize start and stop dates for the observation window
-        if self.start is not None and self.stop is not None:
-            self.start, self.stop = self._normalize_dates(
-                self.start,
-                self.stop,
-            )
-        # If not, self.start and self.stop remain None
-
-        # Catalog objects are lazy-loaded to improve performance
-        self._local_messier = None
-        self._local_planets = None
-        self._local_ngc = None
-        self._local_stars = None
-        self._weather_analysis = None
-
-        # Compute time limit for observation
-        if self.start is not None:
-            if self.conditions.max_return:
-                parts = [int(v) for v in self.conditions.max_return.split(":")]
-                h = parts[0]
-                m = parts[1] if len(parts) > 1 else 0
-                s = parts[2] if len(parts) > 2 else 0
-                self.time_limit = self.start.replace(
-                    hour=h, minute=m, second=s, microsecond=0
-                )
-                # Adjust for overnight observations if necessary.
-                if self.time_limit < self.start:
-                    self.time_limit += timedelta(days=1)
-            else:
-                # If max_return is None, default time_limit to dawn (self.stop)
-                self.time_limit = self.stop
-        # If self.start is None, self.time_limit remains None.
-
-        self._plot = None
-
-    @property
-    def plot(self):
-        if self._plot is None:
-            try:
-                self._plot = apts_plot.Plotter(self)
-            except ImportError:
-                # Fallback if dependencies are missing or plotting is disabled
-                self._plot = apts_plot.NullPlotter()
-            except Exception as e:
-                # Fallback for any other initialization error
-                logger.warning(f"Failed to initialize plotter: {e}")
-                self._plot = apts_plot.NullPlotter()
-        return self._plot
-
-    @plot.setter
-    def plot(self, value):
-        self._plot = value
-
-    @property
-    def local_messier(self) -> Messier:
-        if self._local_messier is None:
-            from apts import catalogs
-
-            self._local_messier = Messier(
-                self.place, catalogs, calculation_date=self.effective_date
-            )
-        return self._local_messier
-
-    @property
-    def local_planets(self) -> SolarObjects:
-        if self._local_planets is None:
-            self._local_planets = SolarObjects(
-                self.place, calculation_date=self.effective_date, lazy=True
-            )
-        return self._local_planets
-
-    @property
-    def local_ngc(self) -> NGC:
-        if self._local_ngc is None:
-            from apts import catalogs
-
-            self._local_ngc = NGC(
-                self.place, catalogs, calculation_date=self.effective_date
-            )
-        return self._local_ngc
-
-    @property
-    def local_stars(self) -> Stars:
-        if self._local_stars is None:
-            from apts import catalogs
-
-            self._local_stars = Stars(
-                self.place, catalogs, calculation_date=self.effective_date
-            )
-        return self._local_stars
-
-    def get_visible_messier(
-        self, language: Optional[str] = None, **args
-    ) -> pd.DataFrame:
-        with language_context(language):
-            from apts.i18n import gettext_
-
-            visible = self.local_messier.get_visible(
-                self.conditions,
-                self.start,
-                self.time_limit,
-                limiting_magnitude=self.limiting_magnitude,
-                **args,
-            )
-            if "Type" in visible.columns:
-                visible["Type"] = visible["Type"].apply(gettext_).astype("string")
-            if "Constellation" in visible.columns:
-                visible["Constellation"] = (
-                    visible["Constellation"].apply(gettext_).astype("string")
-                )
-            return visible
-
-    def get_visible_ngc(self, **args) -> pd.DataFrame:
-        return self.local_ngc.get_visible(
-            self.conditions,
-            self.start,
-            self.time_limit,
-            limiting_magnitude=self.limiting_magnitude,
-            **args,
-        )
-
-    def get_visible_planets(
-        self, language: Optional[str] = None, **args
-    ) -> pd.DataFrame:
-        with language_context(language):
-            from apts.i18n import gettext_
-
-            visible = self.local_planets.get_visible(
-                self.conditions,
-                self.start,
-                self.time_limit,
-                limiting_magnitude=self.limiting_magnitude,
-                **args,
-            )
-            if "Name" in visible.columns:
-                visible["Name"] = visible["Name"].apply(gettext_).astype("string")
-            return visible
-
-    def get_astronomical_events(
-        self,
-        start_date: Optional[datetime] = None,
-        end_date: Optional[datetime] = None,
-        events_to_calculate: Optional[List[EventType]] = None,
-    ):
-        if start_date is None:
-            start_date = self.start
-        if end_date is None:
-            end_date = self.stop
-
-        if start_date is None:
-            start_date = datetime.now(utc)
-        if end_date is None:
-            end_date = start_date + timedelta(days=365)
-
-        events = AstronomicalEvents(
-            self.place, start_date, end_date, events_to_calculate=events_to_calculate
-        )
-        return events.get_events()
-
-    def plot_visible_planets_svg(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.visible_planets_svg(
-                dark_mode_override=dark_mode_override, **args
-            )
-
-    def plot_visible_planets(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.visible_planets(
-                dark_mode_override=dark_mode_override, **args
-            )
-
-    def _normalize_dates(self, start, stop):
-        # If the stop time is earlier than the start time, it means the observation
-        # spans across midnight, so we add one day to the stop time.
-        if stop < start:
-            stop += timedelta(days=1)
-        return (start, stop)
-
-    def plot_messier(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.messier(dark_mode_override=dark_mode_override, **args)
-
-    def plot_planets(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.planets(dark_mode_override=dark_mode_override, **args)
-
-    def _compute_weather_goodness(self, conditions: Optional[Conditions] = None):
-        analysis = self.get_weather_analysis(conditions=conditions)
-        if not analysis:
-            return 0
-
-        good_hours = sum(1 for hour in analysis if hour["is_good_hour"])
-        all_hours = len(analysis)
-
-        logger.debug("Good hours: {} and all hours: {}".format(good_hours, all_hours))
-        if all_hours == 0:
-            return 0
-
-        return good_hours / all_hours * 100
-
-    def _is_moon_condition_met(self, conditions: Conditions) -> bool:
-        if not self.start or not self.stop:
-            return True
-
-        # Use effective_date for moon illumination if available, otherwise fallback to place.date.
-        # We avoid using 'or' here because Skyfield Time objects can raise TypeError when evaluated in boolean context.
-        date_for_illumination = (
-            self.effective_date if self.effective_date is not None else self.place.date
-        )
-        moon_illumination = get_moon_illumination(date_for_illumination)
-        if moon_illumination <= conditions.max_moon_illumination:
-            return True
-
-        # Illumination is too high, check if moon is up during observation window
-        ts = self.place.ts
-        t0 = ts.from_datetime(self.start)
-        t1 = ts.from_datetime(self.stop)
-        # Check at 10 points during the observation
-        times = ts.linspace(t0, t1, 10)
-        alt, _, _ = (
-            self.place.observer.at(times).observe(self.place.moon).apparent().altaz()
-        )
-
-        # Calculate how much of the time the moon is up
-        moon_up_ratio = sum(1 for a in alt.degrees if a > 0) / 10.0
-        # If moon is up for more than (100 - min_weather_goodness)% of the time, it's bad
-        allowed_bad_ratio = (100 - conditions.min_weather_goodness) / 100.0
-
-        if moon_up_ratio > allowed_bad_ratio:
-            logger.info(
-                f"Moon condition not met: illumination {moon_illumination:.1f}% "
-                f"exceeds {conditions.max_moon_illumination}% and moon is up for "
-                f"{moon_up_ratio * 100:.1f}% of the observation window."
-            )
-            return False
-
-        return True
-
-    def is_weather_good(
-        self,
-        conditions: Optional[Conditions] = None,
-        provider_name: Optional[str] = None,
-        force: bool = False,
-    ):
-        effective_conditions = conditions or self.conditions
-        if not force and not self._is_moon_condition_met(effective_conditions):
-            return False
-
-        if self.place.weather is None:
-            logger.info(
-                "is_weather_good: self.place.weather is None, calling get_weather."
-            )
-            self.place.get_weather(
-                provider_name=provider_name,
-                conditions=effective_conditions,
-                observation_window=(self.start, self.stop),
-                force=force,
-            )
-        else:
-            logger.info("is_weather_good: self.place.weather already exists.")
-
-        return (
-            self._compute_weather_goodness(conditions=conditions)
-            >= effective_conditions.min_weather_goodness
-        )
-
-    def plot_weather(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.weather(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_clouds(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.clouds(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_clouds_summary(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.clouds_summary(
-                dark_mode_override=dark_mode_override, **args
-            )
-
-    def plot_precipitation(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.precipitation(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_precipitation_type_summary(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.precipitation_type_summary(
-                dark_mode_override=dark_mode_override, **args
-            )
-
-    def plot_temperature(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.temperature(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_wind(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.wind(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_pressure_and_ozone(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.pressure_and_ozone(
-                dark_mode_override=dark_mode_override, **args
-            )
-
-    def plot_visibility(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.visibility(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_moon_illumination(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.moon_illumination(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_fog(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.fog(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_aurora(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.aurora(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_seeing(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.seeing(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_sqm(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.sqm(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_weather_summary(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.weather_summary(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def to_html(
-        self,
-        custom_template=None,
-        css=None,
-        language: Optional[str] = None,
-    ):
-        with language_context(language):
-            if custom_template:
-                # Security: restrict to allowed template extensions and prevent path traversal.
-                # We allow absolute paths to support templates located outside the package (and for tests),
-                # but we explicitly block directory traversal via '..'.
-                str_template = str(custom_template)
-                if ".." in str_template:
-                    raise ValueError("Path traversal detected in custom template path.")
-
-                if not str_template.lower().endswith((".template", ".html", ".htm")):
-                    raise ValueError(
-                        "Only .template, .html, or .htm files are allowed as custom templates."
-                    )
-                with open(custom_template, "r", encoding="utf-8") as f:
-                    template_content = f.read()
-            else:
-                template_content = self.NOTIFICATION_TEMPLATE.read_text(
-                    encoding="utf-8"
-                )
-
-            if css:
-                # Sanitize CSS to prevent breaking out of <style> block.
-                # Regex handles variations like </style >, </style/>, etc.
-                sanitized_css = re.sub(
-                    r"</style\s*/?>", "", str(css), flags=re.IGNORECASE
-                )
-                # Escape '$' to '$$' to prevent string.Template from interpreting it as a variable.
-                sanitized_css = sanitized_css.replace("$", "$$")
-                style_end_pos = template_content.find("</style>")
-                if style_end_pos != -1:
-                    template_content = (
-                        template_content[:style_end_pos]
-                        + sanitized_css
-                        + template_content[style_end_pos:]
-                    )
-            template = Template(template_content)
-            hourly_weather = self.get_hourly_weather_analysis(language=language)
-            # Sanitize hourly_weather contents
-            sanitized_hourly_weather = []
-            for hour in hourly_weather:
-                sanitized_hour = {}
-                for k, v in hour.items():
-                    if isinstance(v, str):
-                        sanitized_hour[k] = html.escape(v)
-                    elif isinstance(v, list):
-                        sanitized_hour[k] = [
-                            html.escape(item) if isinstance(item, str) else item
-                            for item in v
-                        ]
-                    else:
-                        sanitized_hour[k] = v
-                sanitized_hourly_weather.append(sanitized_hour)
-
-            visible_planets_df = self.get_visible_planets(language=language)
-            messier_df = self.get_visible_messier(language=language)
-
-            data = {
-                "title": "APTS",
-                "start": Utils.format_date(self.start),
-                "stop": Utils.format_date(self.stop),
-                "planets_count": len(visible_planets_df),
-                "messier_count": len(messier_df),
-                "planets_table": visible_planets_df.drop(
-                    columns=["TechnicalName"]
-                ).to_html()
-                if "TechnicalName" in visible_planets_df.columns
-                else visible_planets_df.to_html(),
-                "messier_table": messier_df.to_html(),
-                "equipment_table": self.equipment.data().to_html(),
-                "place_name": html.escape(self.place.name),
-                "lat": np.rad2deg(self.place.lat),
-                "lon": np.rad2deg(self.place.lon),
-                "hourly_weather": sanitized_hourly_weather,
-                "timezone": html.escape(str(self.place.local_timezone)),
-            }
-            return str(template.substitute(data))
-
-    def plot_sun_and_moon_path(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.sun_and_moon_path(
-                dark_mode_override=dark_mode_override, **args
-            )
-
-    def plot_skymap(
-        self,
-        target_name: Optional[str] = None,
-        dark_mode_override: Optional[bool] = None,
-        zoom_deg: Optional[float] = None,
-        star_magnitude_limit: Optional[float] = None,
-        plot_stars: bool = True,
-        plot_messier: bool = False,
-        plot_ngc: bool = False,
-        plot_planets: bool = False,
-        plot_sun: bool = False,
-        plot_moon: bool = False,
-        plot_date: Optional[datetime] = None,
-        flip_horizontally: Optional[bool] = None,
-        flip_vertically: Optional[bool] = None,
-        coordinate_system: Optional[CoordinateSystem] = None,
-        language: Optional[str] = None,
-        texture_mode: bool = False,
-        plot_labels: Optional[bool] = None,
-        **kwargs,
-    ) -> "matplotlib.figure.Figure":
-        """
-        Generates and displays a skymap for a specified celestial object.
-
-        This method provides a high-level interface to the plotting capabilities in `apts.plot`.
-        It can generate either a full-sky polar plot or a zoomed-in Cartesian plot for a
-        given target. The orientation of the view can be flipped to match the output of
-        certain telescopes.
-
-        Args:
-            target_name (Optional[str]): The name of the target object to center the skymap on.
-            dark_mode_override (Optional[bool]): If set, overrides the system's dark mode setting.
-            zoom_deg (Optional[float]): The diameter of the zoomed-in view in degrees. If None, a full skymap is generated.
-            star_magnitude_limit (Optional[float]): The faintest magnitude of stars to plot.
-            plot_stars (bool): Whether to plot stars on the skymap. Defaults to True.
-            plot_messier (bool): Whether to plot Messier objects. Defaults to False.
-            plot_ngc (bool): Whether to plot NGC objects. Defaults to False.
-            plot_planets (bool): Whether to plot planets. Defaults to False.
-            plot_sun (bool): Whether to plot Sun. Defaults to False.
-            plot_moon (bool): Whether to plot Moon. Defaults to False.
-            plot_date (Optional[datetime]): The specific date and time for which to generate the skymap.
-                                            If None, the middle of the observation window is used.
-            flip_horizontally (Optional[bool]): If True, the skymap's horizontal axis is inverted.
-            flip_vertically (Optional[bool]): If True, the skymap's vertical axis is inverted.
-            coordinate_system (Optional[CoordinateSystem]): The coordinate system to use for the skymap (e.g., 'altaz', 'radec').
-            texture_mode (bool): If True, generates a 2D texture of the whole sky.
-            plot_labels (Optional[bool]): Whether to plot labels. Defaults to True (False in texture mode).
-            **kwargs: Additional keyword arguments to pass to the plotting function, including `equipment_id` and `figsize`.
-
-        Returns:
-            A matplotlib Figure object representing the skymap.
-        """
-        with language_context(language):
-            return cast(
-                "matplotlib.figure.Figure",
-                self.plot.skymap(
-                    target_name=target_name,
-                    dark_mode_override=dark_mode_override,
-                    zoom_deg=zoom_deg,
-                    star_magnitude_limit=star_magnitude_limit,
-                    plot_stars=plot_stars,
-                    plot_messier=plot_messier,
-                    plot_ngc=plot_ngc,
-                    plot_planets=plot_planets,
-                    plot_date=plot_date,
-                    flip_horizontally=flip_horizontally,
-                    flip_vertically=flip_vertically,
-                    coordinate_system=coordinate_system,
-                    texture_mode=texture_mode,
-                    plot_labels=plot_labels,
-                    **kwargs,
-                ),
-            )
-
-    def plot_skymap_texture(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        star_magnitude_limit: Optional[float] = None,
-        plot_stars: bool = True,
-        plot_messier: bool = False,
-        plot_ngc: bool = False,
-        plot_planets: bool = False,
-        plot_sun: bool = False,
-        plot_moon: bool = False,
-        plot_date: Optional[datetime] = None,
-        coordinate_system: Optional[CoordinateSystem] = None,
-        language: Optional[str] = None,
-        plot_labels: bool = False,
-        figsize: tuple = (40, 20),
-        **kwargs,
-    ) -> "matplotlib.figure.Figure":
-        """
-        Generates and displays a high-resolution skymap texture.
-
-        Args:
-            dark_mode_override (Optional[bool]): If set, overrides the system's dark mode setting.
-            star_magnitude_limit (Optional[float]): The faintest magnitude of stars to plot.
-            plot_stars (bool): Whether to plot stars. Defaults to True.
-            plot_messier (bool): Whether to plot Messier objects. Defaults to False.
-            plot_ngc (bool): Whether to plot NGC objects. Defaults to False.
-            plot_planets (bool): Whether to plot planets. Defaults to False.
-            plot_sun (bool): Whether to plot Sun. Defaults to False.
-            plot_moon (bool): Whether to plot Moon. Defaults to False.
-            plot_date (Optional[datetime]): The specific date and time for which to generate the skymap.
-            coordinate_system (Optional[CoordinateSystem]): The coordinate system to use.
-            language (Optional[str]): Language for translation.
-            plot_labels (bool): Whether to plot labels. Defaults to False.
-            figsize (tuple): Figure size for hi-res output. Defaults to (40, 20).
-            **kwargs: Additional keyword arguments.
-
-        Returns:
-            A matplotlib Figure object representing the skymap texture.
-        """
-        with language_context(language):
-            return cast(
-                "matplotlib.figure.Figure",
-                self.plot.skymap_texture(
-                    dark_mode_override=dark_mode_override,
-                    star_magnitude_limit=star_magnitude_limit,
-                    plot_stars=plot_stars,
-                    plot_messier=plot_messier,
-                    plot_ngc=plot_ngc,
-                    plot_planets=plot_planets,
-                    plot_sun=plot_sun,
-                    plot_moon=plot_moon,
-                    plot_date=plot_date,
-                    coordinate_system=coordinate_system,
-                    plot_labels=plot_labels,
-                    figsize=figsize,
-                    **kwargs,
-                ),
-            )
-
-    def __str__(self) -> str:
-        return (
-            f"Observation at {self.place.name} from {self.start} to {self.time_limit}"
-        )
-
-    def get_weather_analysis(
-        self,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        provider_name: Optional[str] = None,
-        force: bool = False,
-    ):
-        if conditions is None and self._weather_analysis is not None:
-            return self._weather_analysis
-
-        effective_conditions = conditions or self.conditions
-        # Force a minimal check for is_moon_condition_met to allow it returning reasons later if needed,
-        # but the current logic is to early-exit.
-
-        if not force and not self._is_moon_condition_met(effective_conditions):
-            logger.info("Skipping weather analysis due to moon condition.")
-            return []
-
-        if self.place.weather is None:
-            self.place.get_weather(
-                provider_name=provider_name,
-                conditions=effective_conditions,
-                observation_window=(self.start, self.stop),
-                force=force,
-            )
-            if not force and not self._is_moon_condition_met(effective_conditions):
-                logger.info("Skipping weather analysis due to moon condition.")
-                return []
-
-            if self.place.weather is None:
-                logger.warning("Weather data unavailable after fetch attempt.")
-                return []
-
-        if not all([self.start, self.stop]):
-            logger.warning("Observation window (start, stop) is not fully defined.")
-            return []
-
-        hourly_data = self.place.weather.get_critical_data(self.start, self.stop)
-        if hourly_data.empty:
-            return []
-
-        # Filter by time limit if defined
-        # Ensure we use aware comparison
-        if self.time_limit is not None:
-            t_limit = pd.Timestamp(self.time_limit)
-            if t_limit.tzinfo is None:
-                t_limit = t_limit.tz_localize(self.place.local_timezone)
-
-            hourly_data = hourly_data[hourly_data.time <= t_limit].copy()
-        else:
-            hourly_data = hourly_data.copy()
-
-        if "fog" not in hourly_data.columns:
-            hourly_data["fog"] = 0
-        if "aurora" not in hourly_data.columns:
-            hourly_data["aurora"] = 0
-        if "precipIntensity" not in hourly_data.columns:
-            hourly_data["precipIntensity"] = 0
-        if "precipProbability" not in hourly_data.columns:
-            hourly_data["precipProbability"] = 0
-        if "cloudCover" not in hourly_data.columns:
-            hourly_data["cloudCover"] = 0
-        if "windSpeed" not in hourly_data.columns:
-            hourly_data["windSpeed"] = 0
-        if "temperature" not in hourly_data.columns:
-            hourly_data["temperature"] = 20
-        if "visibility" not in hourly_data.columns:
-            hourly_data["visibility"] = 20
-        if "moonIllumination" not in hourly_data.columns:
-            hourly_data["moonIllumination"] = 0
-        if "seeing" not in hourly_data.columns:
-            hourly_data["seeing"] = 1.5
-        if "sqm" not in hourly_data.columns:
-            hourly_data["sqm"] = 21.0
-
-        # Calculate moon altitudes exactly for each weather data point if not already cached
-        if (
-            "moon_altitude" in hourly_data.columns
-            and not hourly_data["moon_altitude"].isna().all()
-        ):
-            hourly_data["Altitude"] = hourly_data["moon_altitude"]
-        else:
-            ts = self.place.ts
-            times = ts.from_datetimes(hourly_data["time"].tolist())
-            alt, _, _ = (
-                self.place.observer.at(times)
-                .observe(self.place.moon)
-                .apparent()
-                .altaz()
-            )
-            hourly_data["Altitude"] = alt.degrees
-
-        for col in [
-            "cloudCover",
-            "precipProbability",
-            "precipIntensity",
-            "windSpeed",
-            "temperature",
-            "visibility",
-            "moonIllumination",
-            "fog",
-            "aurora",
-            "seeing",
-            "sqm",
-        ]:
-            if col in hourly_data.columns:
-                hourly_data[col] = pd.to_numeric(hourly_data[col], errors="coerce")
-
-        effective_conditions = conditions or self.conditions
-        with language_context(language):
-            from apts.i18n import gettext_
-
-            # Vectorized condition checks
-            is_bad_clouds = hourly_data.cloudCover.isna() | (
-                hourly_data.cloudCover > effective_conditions.max_clouds
-            )
-            is_bad_precip_prob = hourly_data.precipProbability.isna() | (
-                hourly_data.precipProbability
-                > effective_conditions.max_precipitation_probability
-            )
-            is_bad_precip_intens = hourly_data.precipIntensity.isna() | (
-                hourly_data.precipIntensity
-                > effective_conditions.max_precipitation_intensity
-            )
-            is_bad_wind = hourly_data.windSpeed.isna() | (
-                hourly_data.windSpeed > effective_conditions.max_wind
-            )
-            is_bad_temp = hourly_data.temperature.isna() | (
-                (effective_conditions.min_temperature > hourly_data.temperature)
-                | (hourly_data.temperature > effective_conditions.max_temperature)
-            )
-            is_bad_vis = hourly_data.visibility.isna() | (
-                hourly_data.visibility < effective_conditions.min_visibility
-            )
-            is_bad_fog = hourly_data.fog.isna() | (
-                hourly_data.fog > effective_conditions.max_fog
-            )
-            is_bad_moon = (hourly_data["Altitude"] > 0) & (
-                hourly_data.moonIllumination
-                > effective_conditions.max_moon_illumination
-            )
-            is_bad_aurora = hourly_data.aurora.isna() | (
-                hourly_data.aurora < effective_conditions.min_aurora
-            )
-            is_bad_seeing = hourly_data.seeing.isna() | (
-                hourly_data.seeing > effective_conditions.max_seeing
-            )
-            is_bad_sqm = hourly_data.sqm.isna() | (
-                hourly_data.sqm < effective_conditions.min_sqm
-            )
-
-            # Determine good hours
-            is_good_hour_mask = ~(
-                is_bad_clouds
-                | is_bad_precip_prob
-                | is_bad_precip_intens
-                | is_bad_wind
-                | is_bad_temp
-                | is_bad_vis
-                | is_bad_fog
-                | is_bad_moon
-                | is_bad_aurora
-                | is_bad_seeing
-                | is_bad_sqm
-            )
-
-            # Pre-populate analysis_results with vectorized data
-            hourly_data["is_good_hour"] = is_good_hour_mask
-            # Initialize reasons with empty lists
-            # Note: We use a list of empty lists because objects in Series are slow to update individually
-            reasons_col = [[] for _ in range(len(hourly_data))]
-
-            # Only iterate over "bad" hours to generate reason strings (optimization)
-            bad_indices = np.where(~is_good_hour_mask)[0]
-            if len(bad_indices) > 0:
-                # Convert only bad rows to dict for much faster access than iloc in loop
-                bad_rows = hourly_data.iloc[bad_indices].to_dict("records")
-                for i, row in enumerate(bad_rows):
-                    idx = bad_indices[i]
-                    reasons = []
-                    if is_bad_clouds.iloc[idx]:
-                        reasons.append(
-                            gettext_(
-                                "Cloud cover %(cloud_cover)s%% exceeds limit of %(max_clouds)s%%"
-                            )
-                            % {
-                                "cloud_cover": f"{row['cloudCover']:.1f}",
-                                "max_clouds": effective_conditions.max_clouds,
-                            }
-                        )
-                    if is_bad_precip_prob.iloc[idx]:
-                        reasons.append(
-                            gettext_(
-                                "Precipitation probability %(precip_prob)s%% exceeds limit of %(max_precip_prob)s%%"
-                            )
-                            % {
-                                "precip_prob": f"{row['precipProbability']:.1f}",
-                                "max_precip_prob": effective_conditions.max_precipitation_probability,
-                            }
-                        )
-                    if is_bad_precip_intens.iloc[idx]:
-                        reasons.append(
-                            gettext_(
-                                "Precipitation intensity %(precip_intens)s mm exceeds limit of %(max_precip_intens)s mm"
-                            )
-                            % {
-                                "precip_intens": f"{row['precipIntensity']:.1f}",
-                                "max_precip_intens": effective_conditions.max_precipitation_intensity,
-                            }
-                        )
-                    if is_bad_wind.iloc[idx]:
-                        reasons.append(
-                            gettext_(
-                                "Wind speed %(wind_speed)s km/h exceeds limit of %(max_wind)s km/h"
-                            )
-                            % {
-                                "wind_speed": f"{row['windSpeed']:.1f}",
-                                "max_wind": effective_conditions.max_wind,
-                            }
-                        )
-                    if is_bad_temp.iloc[idx]:
-                        reasons.append(
-                            gettext_(
-                                "Temperature %(temp)s°C out of range (%(min_temp)s - %(max_temp)s°C)"
-                            )
-                            % {
-                                "temp": f"{row['temperature']:.1f}",
-                                "min_temp": effective_conditions.min_temperature,
-                                "max_temp": effective_conditions.max_temperature,
-                            }
-                        )
-                    if is_bad_vis.iloc[idx]:
-                        reasons.append(
-                            gettext_(
-                                "Visibility %(vis)s km below limit of %(min_vis)s km"
-                            )
-                            % {
-                                "vis": f"{row['visibility']:.1f}",
-                                "min_vis": effective_conditions.min_visibility,
-                            }
-                        )
-                    if is_bad_fog.iloc[idx]:
-                        reasons.append(
-                            gettext_("Fog %(fog)s%% exceeds limit of %(max_fog)s%%")
-                            % {
-                                "fog": f"{row['fog']:.1f}",
-                                "max_fog": effective_conditions.max_fog,
-                            }
-                        )
-                    if is_bad_moon.iloc[idx]:
-                        reasons.append(
-                            gettext_(
-                                "Moon illumination %(illum)s%% exceeds limit of %(max_illum)s%% while moon is up"
-                            )
-                            % {
-                                "illum": f"{row['moonIllumination']:.1f}",
-                                "max_illum": effective_conditions.max_moon_illumination,
-                            }
-                        )
-                    if is_bad_aurora.iloc[idx]:
-                        reasons.append(
-                            gettext_(
-                                "Aurora %(aurora)s%% below limit of %(min_aurora)s%%"
-                            )
-                            % {
-                                "aurora": f"{row['aurora']:.1f}",
-                                "min_aurora": effective_conditions.min_aurora,
-                            }
-                        )
-                    if is_bad_seeing.iloc[idx]:
-                        reasons.append(
-                            gettext_(
-                                "Seeing %(seeing)s arcsec exceeds limit of %(max_seeing)s arcsec"
-                            )
-                            % {
-                                "seeing": f"{row['seeing']:.1f}",
-                                "max_seeing": effective_conditions.max_seeing,
-                            }
-                        )
-                    if is_bad_sqm.iloc[idx]:
-                        reasons.append(
-                            gettext_(
-                                "Sky brightness %(sqm)s mag/arcsec² below limit of %(min_sqm)s mag/arcsec²"
-                            )
-                            % {
-                                "sqm": f"{row['sqm']:.1f}",
-                                "min_sqm": effective_conditions.min_sqm,
-                            }
-                        )
-                    reasons_col[idx] = reasons
-
-            hourly_data["reasons"] = reasons_col
-
-            # Rename columns to match expected output dictionary keys
-            rename_map = {
-                "cloudCover": "clouds",
-                "precipProbability": "precipitation",
-                "precipIntensity": "precipitation_intensity",
-                "windSpeed": "wind_speed",
-                "moonIllumination": "moon_illumination",
-            }
-            # Select and rename columns for the final result
-            final_cols = [
-                "time",
-                "is_good_hour",
-                "reasons",
-                "temperature",
-                "cloudCover",
-                "precipProbability",
-                "precipIntensity",
-                "windSpeed",
-                "visibility",
-                "moonIllumination",
-                "fog",
-                "aurora",
-                "seeing",
-                "sqm",
-            ]
-            result_df = hourly_data[final_cols].rename(columns=rename_map)
-            analysis_results = result_df.to_dict("records")
-        if conditions is None:
-            self._weather_analysis = analysis_results
-        return analysis_results
-
-    def get_hourly_weather_analysis(
-        self,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        provider_name: Optional[str] = None,
-        force: bool = False,
-    ):
-        return self.get_weather_analysis(
-            language=language,
-            conditions=conditions,
-            provider_name=provider_name,
-            force=force,
-        )
